OCRT1 and 2 are fixed

+++ tests/TestsORCT.py
@@ -46,7 +46,7 @@ class TestORCT(unittest.TestCase):
 
     def test_orct12(self):
         bayer = cv2.imread("../Data/image.bmp")
-        bayer = np.sum(bayer, axis=2).astype('float64')
+        bayer = np.sum(bayer, axis=2).astype("float32")
         filtered = compute_orct2(compute_orct1(bayer))
 
         self.compressionEvaluation.evaluate(bayer, "before ocrt")
@@ -56,7 +56,7 @@ class TestORCT(unittest.TestCase):
     def test_ocrtWithDataset(self):
         rgbImages = self.datasetUtils.loadFoodDataset()
         cfaImages, image_size = self.datasetUtils.convertDatasetToCFA(rgbImages)
-        bayer = cfaImages[0, :, :, 0]
+        bayer = cfaImages[0, :, :, 0].astype("float32")
         filtered = compute_orct2(compute_orct1(bayer))
 
         self.compressionEvaluation.evaluate(bayer, "before ocrt")
